refactor: split actor out of FilebasedSnapshotStore to allow using in tests with out actors
