+++ zeebe/scheduler/src/main/java/io/camunda/zeebe/scheduler/Actor.java
@@ -10,6 +10,7 @@ package io.camunda.zeebe.scheduler;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.util.Loggers;
 import java.time.Duration;
+import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -123,6 +124,12 @@ public abstract class Actor implements AutoCloseable, AsyncClosable, Concurrency
   public <T> void runOnCompletion(
       final ActorFuture<T> future, final BiConsumer<T, Throwable> callback) {
     actor.runOnCompletion(future, callback);
+  }
+
+  @Override
+  public <T> void runOnCompletion(
+      final Collection<ActorFuture<T>> actorFutures, final Consumer<Throwable> callback) {
+    actor.runOnCompletion(actorFutures, callback);
   }
 
   @Override
