+++ zeebe/snapshot/src/main/java/io/camunda/zeebe/snapshots/impl/FileBasedReceivedSnapshot.java
@@ -7,7 +7,7 @@
  */
 package io.camunda.zeebe.snapshots.impl;
 
-import io.camunda.zeebe.scheduler.ActorControl;
+import io.camunda.zeebe.scheduler.ConcurrencyControl;
 import io.camunda.zeebe.scheduler.future.ActorFuture;
 import io.camunda.zeebe.scheduler.future.CompletableActorFuture;
 import io.camunda.zeebe.snapshots.PersistedSnapshot;
@@ -31,8 +31,8 @@ public class FileBasedReceivedSnapshot implements ReceivedSnapshot {
   private static final int BLOCK_SIZE = 512 * 1024;
 
   private final Path directory;
-  private final ActorControl actor;
-  private final FileBasedSnapshotStore snapshotStore;
+  private final ConcurrencyControl actor;
+  private final FileBasedSnapshotStoreImpl snapshotStore;
 
   private final FileBasedSnapshotId snapshotId;
   private long expectedSnapshotChecksum;
@@ -45,8 +45,8 @@ public class FileBasedReceivedSnapshot implements ReceivedSnapshot {
   FileBasedReceivedSnapshot(
       final FileBasedSnapshotId snapshotId,
       final Path directory,
-      final FileBasedSnapshotStore snapshotStore,
-      final ActorControl actor) {
+      final FileBasedSnapshotStoreImpl snapshotStore,
+      final ConcurrencyControl actor) {
     this.snapshotId = snapshotId;
     this.snapshotStore = snapshotStore;
     this.directory = directory;
@@ -111,7 +111,7 @@ public class FileBasedReceivedSnapshot implements ReceivedSnapshot {
     checksumCollection.updateFromBytes(
         snapshotFile.getFileName().toString(), snapshotChunk.getContent());
 
-    if (snapshotChunk.getChunkName().equals(FileBasedSnapshotStore.METADATA_FILE_NAME)) {
+    if (snapshotChunk.getChunkName().equals(FileBasedSnapshotStoreImpl.METADATA_FILE_NAME)) {
       try {
         collectMetadata(snapshotChunk);
       } catch (final IOException e) {
@@ -228,7 +228,11 @@ public class FileBasedReceivedSnapshot implements ReceivedSnapshot {
   @Override
   public ActorFuture<PersistedSnapshot> persist() {
     final CompletableActorFuture<PersistedSnapshot> future = new CompletableActorFuture<>();
-    actor.call(() -> persistInternal(future));
+    actor.call(
+        () -> {
+          persistInternal(future);
+          return null;
+        });
     return future;
   }
 
@@ -282,7 +286,7 @@ public class FileBasedReceivedSnapshot implements ReceivedSnapshot {
         // backward compatibility
         metadata =
             new FileBasedSnapshotMetadata(
-                FileBasedSnapshotStore.VERSION,
+                FileBasedSnapshotStoreImpl.VERSION,
                 snapshotId.getProcessedPosition(),
                 snapshotId.getExportedPosition(),
                 Long.MAX_VALUE);
@@ -302,8 +306,6 @@ public class FileBasedReceivedSnapshot implements ReceivedSnapshot {
     return "FileBasedReceivedSnapshot{"
         + "directory="
         + directory
-        + ", snapshotStore="
-        + snapshotStore.getName()
         + ", metadata="
         + snapshotId
         + '}';
